load et save OK

+++ src/base/config.h
@@ -1,17 +1,18 @@
 #ifndef CONFIG_H
 #define CONFIG_H
+#include <Arduino.h>
 
 typedef struct data {
-    int val;
-    int min;
-    int max;
-    int buffer;
+    byte val;
+    byte min;
+    byte max;
+    byte buffer;
 } data;
 
 typedef struct algo {
-    int in;
-    int out;
-    int action_num;
+    byte in;
+    byte out;
+    byte action_num;
 } algo;
 
 void init_data();
